refactor: 패키지 구성
- 역할에 맞는 패키지 구성
