+++ src/test/java/racinggame/domain/CarTest.java
@@ -1,9 +1,10 @@
-package racinggame;
+package racinggame.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import org.junit.jupiter.api.Test;
+import racinggame.domain.strategy.MovingStrategy;
 
 class CarTest {
